Eliminación de la herencia de solapamientos
